+++ src/Common/Impl/BTreeCache_Stub.java
@@ -7,7 +7,7 @@ public class BTreeCache_Stub implements Common.BTreeCacheInterface{
 
 
 
-@Override public BTreeNodeInterface searchNode(long nodeID){return null;}
+@Override public BTreeNodeInterface searchNode(int nodeID){return null;}
 
 @Override public void insertNode(BTreeNodeInterface node){}
