+++ src/Common/Impl/BTreeFile_Stub.java
@@ -11,13 +11,13 @@ public class BTreeFile_Stub implements BTreeFileInterface{
 
 @Override public BTreeNodeInterface loadFromFile(String sourceFile) throws OmniException{return null;}
 
-@Override public BTreeNodeInterface readNode(long nodeID) throws OmniException{return null;}
+@Override public BTreeNodeInterface readNode(int nodeID) throws OmniException{return null;}
 
-@Override public long allocateNode() throws OmniException{return 0;}
+@Override public int allocateNode() throws OmniException{return 0;}
 
 @Override public void writeNode(BTreeNodeInterface node) throws OmniException{}
 
-@Override public void setRootNode(long nodeID){}
+@Override public void setRootNode(int nodeID){}
